requirements.txt
aiohttp==3.12.15
apscheduler==3.11.0
discord.py==2.5.2
python-dotenv==1.1.1
prometheus-client==0.22.1
SQLAlchemy==2.0.42
alembic==1.16.4
PyYAML==6.0.2
jsonschema==4.25.0

telethon==1.36.0
black==25.1.0
